split out the tab contents into seperate files
