+++ x-pack/plugins/enterprise_search/public/applications/analytics/components/analytics_collection_view/analytics_collection_integrate/analytics_collection_integrate_javascript_embed.tsx
@@ -1,0 +1,103 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import React from 'react';
+
+import { EuiCodeBlock, EuiLink, EuiText } from '@elastic/eui';
+
+import { i18n } from '@kbn/i18n';
+import { FormattedMessage } from '@kbn/i18n-react';
+
+export const javascriptEmbedSteps = (webClientSrc: string, analyticsDNSUrl: string) => [
+  {
+    title: i18n.translate(
+      'xpack.enterpriseSearch.analytics.collections.collectionsView.integrateTab.javascriptEmbed.stepOne.title',
+      {
+        defaultMessage: 'Embed onto site',
+      }
+    ),
+    children: (
+      <>
+        <EuiText grow={false}>
+          <p>
+            {i18n.translate(
+              'xpack.enterpriseSearch.analytics.collections.collectionsView.integrateTab.javascriptEmbed.stepOne.description',
+              {
+                defaultMessage:
+                  'Embed the behavioral analytics JavaScript snippet on every page of the website or application you’d like to track.',
+              }
+            )}
+          </p>
+          <EuiCodeBlock language="html" isCopyable>
+            {`<script src="${webClientSrc}" data-dsn="${analyticsDNSUrl}"></script>`}
+          </EuiCodeBlock>
+        </EuiText>
+      </>
+    ),
+  },
+  {
+    title: i18n.translate(
+      'xpack.enterpriseSearch.analytics.collections.collectionsView.integrateTab.javascriptEmbed.stepTwo.title',
+      {
+        defaultMessage: 'Initialise the client',
+      }
+    ),
+    children: (
+      <>
+        <EuiText grow={false}>
+          <p>
+            {i18n.translate(
+              'xpack.enterpriseSearch.analytics.collections.collectionsView.integrateTab.javascriptEmbed.stepTwo.description',
+              {
+                defaultMessage:
+                  'You must initialise the client before you can track events. We recommend initialising just below the script tag.',
+              }
+            )}
+          </p>
+          <EuiCodeBlock language="html" isCopyable>
+            {'<script type="text/javascript">window.elasticAnalytics.createTracker();</script>'}
+          </EuiCodeBlock>
+        </EuiText>
+      </>
+    ),
+  },
+  {
+    title: i18n.translate(
+      'xpack.enterpriseSearch.analytics.collections.collectionsView.integrateTab.searchuiEmbed.stepThree.title',
+      {
+        defaultMessage: 'Track individual events',
+      }
+    ),
+    children: (
+      <>
+        <EuiText grow={false}>
+          <p>
+            <FormattedMessage
+              id="xpack.enterpriseSearch.analytics.collections.collectionsView.integrateTab.javascriptEmbed.stepThree.description"
+              defaultMessage="Track individual events, like clicks, by calling the trackEvent method. {link}"
+              values={{
+                link: (
+                  <EuiLink href="http://www.elastic.co" target="_blank">
+                    {i18n.translate(
+                      'xpack.enterpriseSearch.analytics.collections.collectionsView.integrateTab.javascriptEmbed.stepThree.link',
+                      {
+                        defaultMessage: 'Learn more about tracking events',
+                      }
+                    )}
+                  </EuiLink>
+                ),
+              }}
+            />
+          </p>
+          <EuiCodeBlock language="html" isCopyable>
+            {`<script src="${webClientSrc}" data-dsn="${analyticsDNSUrl}" defer></script>`}
+          </EuiCodeBlock>
+        </EuiText>
+      </>
+    ),
+  },
+];
